Create UniqueImages.java
Consider this bad program. It is trying to print the number of unique images. For the purposes of this program, two images are considered equal if the filename is the same (disregarding the file extension), and the same number of pixels. However, this program has a serious bug. Please find the bug, fix it, and submit the corrected program.